Refactored TileService into an extenstion method
