+++ src/PDGToolkitCore/Application/Extensions/TileExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using PDGToolkitCore.Domain.Models;
 
 namespace PDGToolkitCore.Application.Extensions
@@ -20,6 +21,20 @@ namespace PDGToolkitCore.Application.Extensions
             }
             
             tiles.Add(replacement);
+        }
+        
+        public static bool HasTwoAdjacentFloorTiles(this Tile tile, List<Tile> allTiles)
+        {
+            var up = new Position(tile.Position.X, tile.Position.Y + 1);
+            var down = new Position(tile.Position.X, tile.Position.Y - 1);
+            var left = new Position(tile.Position.X - 1, tile.Position.Y);
+            var right = new Position(tile.Position.X + 1, tile.Position.Y);
+            var adjacentFloorTiles = allTiles.Where(t => t.Type.Equals(TileType.Floor)).ToList().FindAll(t =>
+                t.Position.Equals(up) || t.Position.Equals(down) || t.Position.Equals(left) ||
+                t.Position.Equals(right));
+
+            var adjacentFloorPositions = adjacentFloorTiles.Select(a => a.Position).ToList(); 
+            return adjacentFloorPositions.Contains(up) && adjacentFloorPositions.Contains(down) || adjacentFloorPositions.Contains(right) && adjacentFloorPositions.Contains(left);
         }
     }
 }
