+++ tests/PDGToolkitCore.UnitTests/TileExtensionsTests.cs
@@ -10,7 +10,7 @@ namespace PDGToolkitCore.UnitTests
     [TestFixture]
     public class TileExtensionsTests
     {
-        private readonly List<Tile> allTiles = new List<Tile>();
+        private readonly List<Tile> tilesWithSharedPositions = new List<Tile>();
         private readonly Position sharedPosition = new Position(0,0);
         
         [SetUp]
@@ -18,19 +18,131 @@ namespace PDGToolkitCore.UnitTests
         {
             for (var i = 0; i < 10; i++)
             {
-                allTiles.Add(new Tile(TileType.Wall, sharedPosition));
+                tilesWithSharedPositions.Add(new Tile(TileType.Wall, sharedPosition));
             }
         }
 
         [Test]
-        public void Extension_ReplacesTilesWithSharedPosition()
+        public void TileExtensions_ReplacesTilesWithSharedPosition()
         {
             var replacement = new Tile(TileType.Floor, sharedPosition);
             
-            allTiles.ReplaceTilesWithOtherTile(replacement);
-            allTiles.Count().Should().Be(1);
-            allTiles.First().Should().BeEquivalentTo(replacement);
+            tilesWithSharedPositions.ReplaceTilesWithOtherTile(replacement);
+            tilesWithSharedPositions.Count().Should().Be(1);
+            tilesWithSharedPositions.First().Should().BeEquivalentTo(replacement);
         }
         
+        [Test]
+        public void TileExtensions_HasTwoVerticallyAdjacentFloorTilesTest()
+        {
+            /* create 3 horizontal lines, composed of 10 tiles
+            like so:    FFFFF
+                        WWWWW
+                        FFFFF
+            where F = Floor tile, W = Wall tile
+             */
+            var allTiles = TileBuilder.Create()
+                .WithHorizontalLineOfFloorsStartingAtPosition(5, new Position(0, 0))
+                .WithHorizontalWallStartingAtPosition(5, new Position(0, 1))
+                .WithHorizontalLineOfFloorsStartingAtPosition(5, new Position(0, 2))
+                .Build();
+            
+            var aWallTile = new Tile(TileType.Wall, new Position(0,1));
+
+            var result = aWallTile.HasTwoAdjacentFloorTiles(allTiles);
+            result.Should().BeTrue();
+        }
+        
+        [Test]
+        public void TileExtensions_HasNoVerticallyAdjacentFloorTilesTest()
+        {
+            /* create 3 horizontal lines, composed of 10 tiles
+            like so:    FFFFF
+                        WWWWW
+                        WWWWW
+            where F = Floor tile, W = Wall tile
+             */
+            var allTiles = TileBuilder.Create()
+                .WithHorizontalLineOfFloorsStartingAtPosition(5, new Position(0, 0))
+                .WithHorizontalWallStartingAtPosition(5, new Position(0, 1))
+                .WithHorizontalWallStartingAtPosition(5, new Position(0, 2))
+                .Build();
+            
+            var aWallTile = new Tile(TileType.Wall, new Position(0,1));
+
+            var result = aWallTile.HasTwoAdjacentFloorTiles(allTiles);
+            result.Should().BeFalse();
+        }
+        
+        [Test]
+        public void TileExtensions_HasTwoHorizontallyAdjacentFloorTilesTest()
+        {
+            /* create 3 vertical lines, composed of 5 tiles
+            like so:    FWF
+                        FWF
+                        FWF
+                        FWF
+                        FWF
+            where F = Floor tile, W = Wall tile
+             */
+            var allTiles = TileBuilder.Create()
+                .WithVerticalLineOfFloorsStartingAtPosition(5, new Position(0, 0))
+                .WithVerticalWallStartingAtPosition(5, new Position(1, 0))
+                .WithVerticalLineOfFloorsStartingAtPosition(5, new Position(2, 0))
+                .Build();
+            
+            var aWallTile = new Tile(TileType.Wall, new Position(1,0));
+
+            var result = aWallTile.HasTwoAdjacentFloorTiles(allTiles);
+            result.Should().BeTrue();
+        }
+        
+        [Test]
+        public void TileExtensions_HasNoHorizontallyAdjacentFloorTilesTest()
+        {
+            /* create 3 vertical lines, composed of 5 tiles
+            like so:    FWW
+                        FWW
+                        FWW
+                        FWW
+                        FWW
+            where F = Floor tile, W = Wall tile
+             */
+            var allTiles = TileBuilder.Create()
+                .WithVerticalLineOfFloorsStartingAtPosition(5, new Position(0, 0))
+                .WithVerticalWallStartingAtPosition(5, new Position(1, 0))
+                .WithVerticalWallStartingAtPosition(5, new Position(2, 0))
+                .Build();
+            
+            var aWallTile = new Tile(TileType.Wall, new Position(1,0));
+
+            var result = aWallTile.HasTwoAdjacentFloorTiles(allTiles);
+            result.Should().BeFalse();
+        }
+
+        [Test] public void TileExtensions_HasBothVerticallyAndHorizontallyAdjacentFloorTiles()
+        {
+            /* Given a set of tiles with positions
+            like so:    012 x
+            0            F
+            1           FWF
+            2            F
+            y
+            where F = Floor tile, W = Wall tile
+             */
+            var allTiles = new List<Tile>
+            {
+                new Tile(TileType.Floor, new Position(1,0)),
+                new Tile(TileType.Floor, new Position(0,1)),
+                new Tile(TileType.Wall, new Position(1,1)),
+                new Tile(TileType.Floor, new Position(2,1)),
+                new Tile(TileType.Floor, new Position(1,2))
+            };
+            
+            var aWallTile = new Tile(TileType.Wall, new Position(1,1));
+
+            var result = aWallTile.HasTwoAdjacentFloorTiles(allTiles);
+            result.Should().BeTrue();
+        }
     }
 }
